requirements.txt
tensorflow>=2.15.0
tensorflow-hub>=0.15.0
tensorflow-text>=2.15.0
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.13.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
